Update rediscommand.cpp
Changed to int64 to fit full range of redis accepted input.

+++ common/rediscommand.cpp
@@ -89,7 +89,7 @@ void RedisCommand::formatHDEL(const std::string& key, const std::vector<std::str
 }
 
 /* Format EXPIRE key field command */
-void RedisCommand::formatEXPIRE(const std::string& key, const int& ttl)
+void RedisCommand::formatEXPIRE(const std::string& key, const int64_t& ttl)
 {
     return format("EXPIRE %s %d", key.c_str(), ttl);
 }
